missing documentations added

+++ SuitSupplyProductTest.Data/ProductDataAccess.cs
@@ -5,6 +5,9 @@ using System.Linq;
 
 namespace SuitSupplyProductTest.Data
 {
+    /// <summary>
+    /// EF DB access for Product
+    /// </summary>
     public class ProductDataAccess : IProductDataAccess
     {
         private readonly EfDbContext context;
@@ -14,6 +17,11 @@ namespace SuitSupplyProductTest.Data
             this.context = context;
         }
 
+        /// <summary>
+        /// Add new product to database
+        /// </summary>
+        /// <param name="product">Product to add </param>
+        /// <returns>Product saved to data source</returns>
         public Product InsertProduct(Product product)
         {
             if (product is null)
@@ -26,6 +34,10 @@ namespace SuitSupplyProductTest.Data
             return entity.Entity;
         }
 
+        /// <summary>
+        /// Update a product in the database
+        /// </summary>
+        /// <param name="product">Product to update</param>
         public void UpdateProduct(Product product)
         {
             if (product is null)
@@ -35,30 +47,44 @@ namespace SuitSupplyProductTest.Data
             context.Entry(product).State = EntityState.Modified;
         }
 
+        /// <summary>
+        /// Get a product by Id
+        /// </summary>
+        /// <param name="id">ID</param>
+        /// <returns>Product</returns>
         public Product GetProductById(int id)
         {
             return context.Products.Find(id);
         }
 
+        /// <summary>
+        /// Get a product by code
+        /// </summary>
+        /// <param name="code">Code</param>
+        /// <returns>Product</returns>
         public Product GetProductByCode(string code)
         {
             return context.Products.FirstOrDefault(p => p.Code == code);
         }
 
+        /// <summary>
+        /// Get all products in the database
+        /// </summary>
+        /// <returns>Products</returns>
         public IEnumerable<Product> GetProducts()
         {
             return context.Products.AsEnumerable();
         }
 
-        public void DeleteProduct(int productId)
-        {
-            var product = context.Products.Find(productId);
-            context.Products.Remove(product);
-            context.SaveChanges();
-        }
-
+        /// <summary>
+        /// Delete a product from database
+        /// </summary>
+        /// <param name="product">Product to delete</param>
         public void DeleteProduct(Product product)
         {
+            if (product is null)
+                throw new ArgumentNullException("product");
+
             context.Products.Remove(product);
             context.SaveChanges();
         }
